Remove warnings HiddenVolumeActive Start of implementation of new debugging features

+++ stick20responsedialog.h
@@ -32,9 +32,9 @@
 extern "C" {
 #endif
     extern char DebugText_Stick20[STICK20_DEBUG_TEXT_LEN];
-    extern unsigned long DebugTextlen_Stick20;
-    extern int DebugingActive;
-    extern int DebugingStick20PoolingActive;
+    extern int  DebugTextlen_Stick20;
+    extern int  DebugingActive;
+    extern int  DebugingStick20PoolingActive;
 
     void DebugAppendText (char *Text);
 #ifdef __cplusplus
